consolidate validation methods
